requirements.txt
json_repair==0.30.0
datasketch==1.6.5
colorama==0.4.6
py-solc-x==2.0.3
solc_select==1.0.4
slither-analyzer==0.10.4
google-generativeai==0.8.3
networkx==3.0
scipy==1.10.1
torch==2.0.1
transformers==4.44.0
deepspeed==0.15.2
accelerate==1.0.1
peft==0.5.0
numpy==1.20.0
jinja2==3.1.3
flash-attn==2.5.6
datasets==2.18.0
modelscope==1.13.3
pydantic==2.9.2
bitsandbytes==0.43.0
